Refactor the elliptic curve cryptography classes
