--- bitcoin/constants.py
@@ -1,7 +1,0 @@
-from .s256_field import S256Field
-from .s256_point import S256Point, G
-
-A = S256Point.A
-B = S256Point.B
-N = S256Point.N
-P = S256Field.P
